test/requirements.txt
Flask==2.3.2
tflite-runtime==2.5.0
numpy
Pillow
